UnconvUserService: Split into Interface and an implementation
